feat(block): implement based sequencing with enhanced block management
Add support for based sequencing with transaction ordering from base layer
Implement batch transaction handling in block creation process
Add empty block creation for chain liveness when no transactions available
Enhance logging for better operational visibility
Improve error handling during batch processing and block submission
Update documentation for block creation and DA inclusion processes
Refactor code for better readability and maintainability
