+++ block/manager.go
@@ -4,10 +4,8 @@ import (
 	"bytes"
 	"context"
 	"encoding/binary"
-	"encoding/hex"
 	"errors"
 	"fmt"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -21,8 +19,6 @@ import (
 	ds "github.com/ipfs/go-datastore"
 	"github.com/libp2p/go-libp2p/core/crypto"
 
-	"github.com/rollkit/go-sequencing"
-
 	"github.com/rollkit/rollkit/config"
 	coreda "github.com/rollkit/rollkit/core/da"
 	coreexecutor "github.com/rollkit/rollkit/core/execution"
@@ -30,7 +26,6 @@ import (
 	"github.com/rollkit/rollkit/store"
 	"github.com/rollkit/rollkit/third_party/log"
 	"github.com/rollkit/rollkit/types"
-	rollkitproto "github.com/rollkit/rollkit/types/pb/rollkit"
 )
 
 const (
@@ -476,177 +471,6 @@ func (m *Manager) GetExecutor() coreexecutor.Executor {
 	return m.exec
 }
 
-// BatchRetrieveLoop is responsible for retrieving batches from the sequencer.
-func (m *Manager) BatchRetrieveLoop(ctx context.Context) {
-	m.logger.Info("Starting BatchRetrieveLoop")
-	batchTimer := time.NewTimer(0)
-	defer batchTimer.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-batchTimer.C:
-			start := time.Now()
-			m.logger.Debug("Attempting to retrieve next batch",
-				"chainID", m.genesis.ChainID,
-				"lastBatchHash", hex.EncodeToString(m.lastBatchHash))
-
-			req := coresequencer.GetNextBatchRequest{
-				RollupId:      []byte(m.genesis.ChainID),
-				LastBatchHash: m.lastBatchHash,
-			}
-
-			res, err := m.sequencer.GetNextBatch(ctx, req)
-			if err != nil {
-				m.logger.Error("error while retrieving batch", "error", err)
-				// Always reset timer on error
-				batchTimer.Reset(m.config.Node.BlockTime)
-				continue
-			}
-
-			if res != nil && res.Batch != nil {
-				m.logger.Debug("Retrieved batch",
-					"txCount", len(res.Batch.Transactions),
-					"timestamp", res.Timestamp)
-
-				if h, err := res.Batch.Hash(); err == nil {
-					m.bq.AddBatch(BatchWithTime{Batch: res.Batch, Time: res.Timestamp})
-					if len(res.Batch.Transactions) != 0 {
-						if err := m.store.SetMetadata(ctx, LastBatchHashKey, h); err != nil {
-							m.logger.Error("error while setting last batch hash", "error", err)
-						}
-						m.lastBatchHash = h
-					}
-				}
-			} else {
-				m.logger.Debug("No batch available")
-			}
-
-			// Always reset timer
-			elapsed := time.Since(start)
-			remainingSleep := m.config.Node.BlockTime - elapsed
-			if remainingSleep < 0 {
-				remainingSleep = 0
-			}
-			batchTimer.Reset(remainingSleep)
-		}
-	}
-}
-
-// AggregationLoop is responsible for aggregating transactions into rollup-blocks.
-func (m *Manager) AggregationLoop(ctx context.Context) {
-	initialHeight := m.genesis.InitialHeight //nolint:gosec
-	height := m.store.Height()
-	var delay time.Duration
-
-	// TODO(tzdybal): double-check when https://github.com/celestiaorg/rollmint/issues/699 is resolved
-	if height < initialHeight {
-		delay = time.Until(m.genesis.GenesisTime)
-	} else {
-		lastBlockTime := m.getLastBlockTime()
-		delay = time.Until(lastBlockTime.Add(m.config.Node.BlockTime))
-	}
-
-	if delay > 0 {
-		m.logger.Info("Waiting to produce block", "delay", delay)
-		time.Sleep(delay)
-	}
-
-	// blockTimer is used to signal when to build a block based on the
-	// rollup block time. A timer is used so that the time to build a block
-	// can be taken into account.
-	blockTimer := time.NewTimer(0)
-	defer blockTimer.Stop()
-
-	// Lazy Aggregator mode.
-	// In Lazy Aggregator mode, blocks are built only when there are
-	// transactions or every LazyBlockTime.
-	if m.config.Node.LazyAggregator {
-		m.lazyAggregationLoop(ctx, blockTimer)
-		return
-	}
-
-	m.normalAggregationLoop(ctx, blockTimer)
-}
-
-func (m *Manager) lazyAggregationLoop(ctx context.Context, blockTimer *time.Timer) {
-	// start is used to track the start time of the block production period
-	start := time.Now()
-	// lazyTimer is used to signal when a block should be built in
-	// lazy mode to signal that the chain is still live during long
-	// periods of inactivity.
-	lazyTimer := time.NewTimer(0)
-	defer lazyTimer.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		// the m.bq.notifyCh channel is signalled when batch becomes available in the batch queue
-		case _, ok := <-m.bq.notifyCh:
-			if ok && !m.buildingBlock {
-				// set the buildingBlock flag to prevent multiple calls to reset the time
-				m.buildingBlock = true
-				// Reset the block timer based on the block time.
-				blockTimer.Reset(m.getRemainingSleep(start))
-			}
-			continue
-		case <-lazyTimer.C:
-		case <-blockTimer.C:
-		}
-		// Define the start time for the block production period
-		start = time.Now()
-		if err := m.publishBlock(ctx); err != nil && ctx.Err() == nil {
-			m.logger.Error("error while publishing block", "error", err)
-		}
-		// unset the buildingBlocks flag
-		m.buildingBlock = false
-		// Reset the lazyTimer to produce a block even if there
-		// are no transactions as a way to signal that the chain
-		// is still live.
-		lazyTimer.Reset(m.getRemainingSleep(start))
-	}
-}
-
-func (m *Manager) normalAggregationLoop(ctx context.Context, blockTimer *time.Timer) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-blockTimer.C:
-			// Define the start time for the block production period
-			start := time.Now()
-			if err := m.publishBlock(ctx); err != nil && ctx.Err() == nil {
-				m.logger.Error("error while publishing block", "error", err)
-			}
-			// Reset the blockTimer to signal the next block production
-			// period based on the block time.
-			blockTimer.Reset(m.getRemainingSleep(start))
-		}
-	}
-}
-
-// HeaderSubmissionLoop is responsible for submitting blocks to the DA layer.
-func (m *Manager) HeaderSubmissionLoop(ctx context.Context) {
-	timer := time.NewTicker(m.config.DA.BlockTime)
-	defer timer.Stop()
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-timer.C:
-		}
-		if m.pendingHeaders.isEmpty() {
-			continue
-		}
-		err := m.submitHeadersToDA(ctx)
-		if err != nil {
-			m.logger.Error("error while submitting block to DA", "error", err)
-		}
-	}
-}
-
 func (m *Manager) handleEmptyDataHash(ctx context.Context, header *types.Header) {
 	headerHeight := header.Height()
 	if bytes.Equal(header.DataHash, dataHashForEmptyTxs) {
@@ -675,642 +499,12 @@ func (m *Manager) handleEmptyDataHash(ctx context.Context, header *types.Header)
 	}
 }
 
-// SyncLoop is responsible for syncing blocks.
-//
-// SyncLoop processes headers gossiped in P2P network to know what's the latest block height,
-// block data is retrieved from DA layer.
-func (m *Manager) SyncLoop(ctx context.Context) {
-	daTicker := time.NewTicker(m.config.DA.BlockTime)
-	defer daTicker.Stop()
-	blockTicker := time.NewTicker(m.config.Node.BlockTime)
-	defer blockTicker.Stop()
-	for {
-		select {
-		case <-daTicker.C:
-			m.sendNonBlockingSignalToRetrieveCh()
-		case <-blockTicker.C:
-			m.sendNonBlockingSignalToHeaderStoreCh()
-			m.sendNonBlockingSignalToDataStoreCh()
-		case headerEvent := <-m.headerInCh:
-			// Only validated headers are sent to headerInCh, so we can safely assume that headerEvent.header is valid
-			header := headerEvent.Header
-			daHeight := headerEvent.DAHeight
-			headerHash := header.Hash().String()
-			headerHeight := header.Height()
-			m.logger.Debug("header retrieved",
-				"height", headerHash,
-				"daHeight", daHeight,
-				"hash", headerHash,
-			)
-			if headerHeight <= m.store.Height() || m.headerCache.isSeen(headerHash) {
-				m.logger.Debug("header already seen", "height", headerHeight, "block hash", headerHash)
-				continue
-			}
-			m.headerCache.setHeader(headerHeight, header)
-
-			m.sendNonBlockingSignalToHeaderStoreCh()
-			m.sendNonBlockingSignalToRetrieveCh()
-
-			// check if the dataHash is dataHashForEmptyTxs
-			// no need to wait for syncing Data, instead prepare now and set
-			// so that trySyncNextBlock can progress
-			m.handleEmptyDataHash(ctx, &header.Header)
-
-			err := m.trySyncNextBlock(ctx, daHeight)
-			if err != nil {
-				m.logger.Info("failed to sync next block", "error", err)
-				continue
-			}
-			m.headerCache.setSeen(headerHash)
-		case dataEvent := <-m.dataInCh:
-			data := dataEvent.Data
-			daHeight := dataEvent.DAHeight
-			dataHash := data.Hash().String()
-			dataHeight := data.Metadata.Height
-			m.logger.Debug("data retrieved",
-				"height", dataHash,
-				"daHeight", daHeight,
-				"hash", dataHash,
-			)
-			if dataHeight <= m.store.Height() || m.dataCache.isSeen(dataHash) {
-				m.logger.Debug("data already seen", "height", dataHeight, "data hash", dataHash)
-				continue
-			}
-			m.dataCache.setData(dataHeight, data)
-
-			m.sendNonBlockingSignalToDataStoreCh()
-
-			err := m.trySyncNextBlock(ctx, daHeight)
-			if err != nil {
-				m.logger.Info("failed to sync next block", "error", err)
-				continue
-			}
-			m.dataCache.setSeen(dataHash)
-		case <-ctx.Done():
-			return
-		}
-	}
-}
-
-func (m *Manager) sendNonBlockingSignalToHeaderStoreCh() {
-	select {
-	case m.headerStoreCh <- struct{}{}:
-	default:
-	}
-}
-
-func (m *Manager) sendNonBlockingSignalToDataStoreCh() {
-	select {
-	case m.dataStoreCh <- struct{}{}:
-	default:
-	}
-}
-
-func (m *Manager) sendNonBlockingSignalToRetrieveCh() {
-	select {
-	case m.retrieveCh <- struct{}{}:
-	default:
-	}
-}
-
-// trySyncNextBlock tries to execute as many blocks as possible from the blockCache.
-//
-//	Note: the blockCache contains only valid blocks that are not yet synced
-//
-// For every block, to be able to apply block at height h, we need to have its Commit. It is contained in block at height h+1.
-// If commit for block h+1 is available, we proceed with sync process, and remove synced block from sync cache.
-func (m *Manager) trySyncNextBlock(ctx context.Context, daHeight uint64) error {
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-		}
-		currentHeight := m.store.Height()
-		h := m.headerCache.getHeader(currentHeight + 1)
-		if h == nil {
-			m.logger.Debug("header not found in cache", "height", currentHeight+1)
-			return nil
-		}
-		d := m.dataCache.getData(currentHeight + 1)
-		if d == nil {
-			m.logger.Debug("data not found in cache", "height", currentHeight+1)
-			return nil
-		}
-
-		hHeight := h.Height()
-		m.logger.Info("Syncing header and data", "height", hHeight)
-		// Validate the received block before applying
-		if err := m.execValidate(m.lastState, h, d); err != nil {
-			return fmt.Errorf("failed to validate block: %w", err)
-		}
-		newState, responses, err := m.applyBlock(ctx, h, d)
-		if err != nil {
-			if ctx.Err() != nil {
-				return err
-			}
-			// if call to applyBlock fails, we halt the node, see https://github.com/cometbft/cometbft/pull/496
-			panic(fmt.Errorf("failed to ApplyBlock: %w", err))
-		}
-		err = m.store.SaveBlockData(ctx, h, d, &h.Signature)
-		if err != nil {
-			return SaveBlockError{err}
-		}
-		_, err = m.execCommit(ctx, newState, h, d, responses)
-		if err != nil {
-			return fmt.Errorf("failed to Commit: %w", err)
-		}
-
-		err = m.store.SaveBlockResponses(ctx, hHeight, responses)
-		if err != nil {
-			return SaveBlockResponsesError{err}
-		}
-
-		// Height gets updated
-		m.store.SetHeight(ctx, hHeight)
-
-		if daHeight > newState.DAHeight {
-			newState.DAHeight = daHeight
-		}
-		err = m.updateState(ctx, newState)
-		if err != nil {
-			m.logger.Error("failed to save updated state", "error", err)
-		}
-		m.headerCache.deleteHeader(currentHeight + 1)
-		m.dataCache.deleteData(currentHeight + 1)
-	}
-}
-
-// HeaderStoreRetrieveLoop is responsible for retrieving headers from the Header Store.
-func (m *Manager) HeaderStoreRetrieveLoop(ctx context.Context) {
-	lastHeaderStoreHeight := uint64(0)
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-m.headerStoreCh:
-		}
-		headerStoreHeight := m.headerStore.Height()
-		if headerStoreHeight > lastHeaderStoreHeight {
-			headers, err := m.getHeadersFromHeaderStore(ctx, lastHeaderStoreHeight+1, headerStoreHeight)
-			if err != nil {
-				m.logger.Error("failed to get headers from Header Store", "lastHeaderHeight", lastHeaderStoreHeight, "headerStoreHeight", headerStoreHeight, "errors", err.Error())
-				continue
-			}
-			daHeight := atomic.LoadUint64(&m.daHeight)
-			for _, header := range headers {
-				// Check for shut down event prior to logging
-				// and sending header to headerInCh. The reason
-				// for checking for the shutdown event
-				// separately is due to the inconsistent nature
-				// of the select statement when multiple cases
-				// are satisfied.
-				select {
-				case <-ctx.Done():
-					return
-				default:
-				}
-				// early validation to reject junk headers
-				if !m.isUsingExpectedCentralizedSequencer(header) {
-					continue
-				}
-				m.logger.Debug("header retrieved from p2p header sync", "headerHeight", header.Height(), "daHeight", daHeight)
-				m.headerInCh <- NewHeaderEvent{header, daHeight}
-			}
-		}
-		lastHeaderStoreHeight = headerStoreHeight
-	}
-}
-
-// DataStoreRetrieveLoop is responsible for retrieving data from the Data Store.
-func (m *Manager) DataStoreRetrieveLoop(ctx context.Context) {
-	lastDataStoreHeight := uint64(0)
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-m.dataStoreCh:
-		}
-		dataStoreHeight := m.dataStore.Height()
-		if dataStoreHeight > lastDataStoreHeight {
-			data, err := m.getDataFromDataStore(ctx, lastDataStoreHeight+1, dataStoreHeight)
-			if err != nil {
-				m.logger.Error("failed to get data from Data Store", "lastDataStoreHeight", lastDataStoreHeight, "dataStoreHeight", dataStoreHeight, "errors", err.Error())
-				continue
-			}
-			daHeight := atomic.LoadUint64(&m.daHeight)
-			for _, d := range data {
-				// Check for shut down event prior to logging
-				// and sending header to dataInCh. The reason
-				// for checking for the shutdown event
-				// separately is due to the inconsistent nature
-				// of the select statement when multiple cases
-				// are satisfied.
-				select {
-				case <-ctx.Done():
-					return
-				default:
-				}
-				//TODO: remove junk if possible
-				m.logger.Debug("data retrieved from p2p data sync", "dataHeight", d.Metadata.Height, "daHeight", daHeight)
-				m.dataInCh <- NewDataEvent{d, daHeight}
-			}
-		}
-		lastDataStoreHeight = dataStoreHeight
-	}
-}
-
-func (m *Manager) getHeadersFromHeaderStore(ctx context.Context, startHeight, endHeight uint64) ([]*types.SignedHeader, error) {
-	if startHeight > endHeight {
-		return nil, fmt.Errorf("startHeight (%d) is greater than endHeight (%d)", startHeight, endHeight)
-	}
-	headers := make([]*types.SignedHeader, endHeight-startHeight+1)
-	for i := startHeight; i <= endHeight; i++ {
-		header, err := m.headerStore.GetByHeight(ctx, i)
-		if err != nil {
-			return nil, err
-		}
-		headers[i-startHeight] = header
-	}
-	return headers, nil
-}
-
-func (m *Manager) getDataFromDataStore(ctx context.Context, startHeight, endHeight uint64) ([]*types.Data, error) {
-	if startHeight > endHeight {
-		return nil, fmt.Errorf("startHeight (%d) is greater than endHeight (%d)", startHeight, endHeight)
-	}
-	data := make([]*types.Data, endHeight-startHeight+1)
-	for i := startHeight; i <= endHeight; i++ {
-		d, err := m.dataStore.GetByHeight(ctx, i)
-		if err != nil {
-			return nil, err
-		}
-		data[i-startHeight] = d
-	}
-	return data, nil
-}
-
-// RetrieveLoop is responsible for interacting with DA layer.
-func (m *Manager) RetrieveLoop(ctx context.Context) {
-	// blockFoundCh is used to track when we successfully found a block so
-	// that we can continue to try and find blocks that are in the next DA height.
-	// This enables syncing faster than the DA block time.
-	headerFoundCh := make(chan struct{}, 1)
-	defer close(headerFoundCh)
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-m.retrieveCh:
-		case <-headerFoundCh:
-		}
-		daHeight := atomic.LoadUint64(&m.daHeight)
-		err := m.processNextDAHeader(ctx)
-		if err != nil && ctx.Err() == nil {
-			// if the requested da height is not yet available, wait silently, otherwise log the error and wait
-			if !strings.Contains(err.Error(), ErrHeightFromFutureStr) {
-				m.logger.Error("failed to retrieve block from DALC", "daHeight", daHeight, "errors", err.Error())
-			}
-			continue
-		}
-		// Signal the headerFoundCh to try and retrieve the next block
-		select {
-		case headerFoundCh <- struct{}{}:
-		default:
-		}
-		atomic.AddUint64(&m.daHeight, 1)
-	}
-}
-
-func (m *Manager) processNextDAHeader(ctx context.Context) error {
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	default:
-	}
-
-	// TODO(tzdybal): extract configuration option
-	maxRetries := 10
-	daHeight := atomic.LoadUint64(&m.daHeight)
-
-	var err error
-	m.logger.Debug("trying to retrieve block from DA", "daHeight", daHeight)
-	for r := 0; r < maxRetries; r++ {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-		}
-		headerResp, fetchErr := m.fetchHeaders(ctx, daHeight)
-		if fetchErr == nil {
-			if headerResp.Code == coreda.StatusNotFound {
-				m.logger.Debug("no header found", "daHeight", daHeight, "reason", headerResp.Message)
-				return nil
-			}
-			m.logger.Debug("retrieved potential headers", "n", len(headerResp.Headers), "daHeight", daHeight)
-			for _, bz := range headerResp.Headers {
-				header := new(types.SignedHeader)
-				// decode the header
-				var headerPb rollkitproto.SignedHeader
-				err := headerPb.Unmarshal(bz)
-				if err != nil {
-					m.logger.Error("failed to unmarshal header", "error", err)
-					continue
-				}
-				err = header.FromProto(&headerPb)
-				if err != nil {
-					m.logger.Error("failed to unmarshal header", "error", err)
-					continue
-				}
-				// early validation to reject junk headers
-				if !m.isUsingExpectedCentralizedSequencer(header) {
-					m.logger.Debug("skipping header from unexpected sequencer",
-						"headerHeight", header.Height(),
-						"headerHash", header.Hash().String())
-					continue
-				}
-				blockHash := header.Hash().String()
-				m.headerCache.setDAIncluded(blockHash)
-				err = m.setDAIncludedHeight(ctx, header.Height())
-				if err != nil {
-					return err
-				}
-				m.logger.Info("block marked as DA included", "blockHeight", header.Height(), "blockHash", blockHash)
-				if !m.headerCache.isSeen(blockHash) {
-					// Check for shut down event prior to logging
-					// and sending block to blockInCh. The reason
-					// for checking for the shutdown event
-					// separately is due to the inconsistent nature
-					// of the select statement when multiple cases
-					// are satisfied.
-					select {
-					case <-ctx.Done():
-						return fmt.Errorf("unable to send block to blockInCh, context done: %w", ctx.Err())
-					default:
-					}
-					m.headerInCh <- NewHeaderEvent{header, daHeight}
-				}
-			}
-			return nil
-		}
-
-		// Track the error
-		err = errors.Join(err, fetchErr)
-		// Delay before retrying
-		select {
-		case <-ctx.Done():
-			return err
-		case <-time.After(100 * time.Millisecond):
-		}
-	}
-	return err
-}
-
-func (m *Manager) isUsingExpectedCentralizedSequencer(header *types.SignedHeader) bool {
-	return bytes.Equal(header.ProposerAddress, m.genesis.ProposerAddress) && header.ValidateBasic() == nil
-}
-
-func (m *Manager) fetchHeaders(ctx context.Context, daHeight uint64) (coreda.ResultRetrieveHeaders, error) {
-	var err error
-	ctx, cancel := context.WithTimeout(ctx, 60*time.Second) //TODO: make this configurable
-	defer cancel()
-	headerRes := m.dalc.RetrieveHeaders(ctx, daHeight)
-	if headerRes.Code == coreda.StatusError {
-		err = fmt.Errorf("failed to retrieve block: %s", headerRes.Message)
-	}
-	return headerRes, err
-}
-
 func (m *Manager) getSignature(header types.Header) (types.Signature, error) {
 	b, err := header.MarshalBinary()
 	if err != nil {
 		return nil, err
 	}
 	return m.proposerKey.Sign(b)
-}
-
-func (m *Manager) getTxsFromBatch() ([][]byte, *time.Time, error) {
-	batch := m.bq.Next()
-	if batch == nil {
-		// batch is nil when there is nothing to process
-		return nil, nil, ErrNoBatch
-	}
-	return batch.Transactions, &batch.Time, nil
-}
-
-func (m *Manager) publishBlock(ctx context.Context) error {
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	default:
-	}
-
-	if !m.isProposer {
-		return ErrNotProposer
-	}
-
-	if m.config.Node.MaxPendingBlocks != 0 && m.pendingHeaders.numPendingHeaders() >= m.config.Node.MaxPendingBlocks {
-		return fmt.Errorf("refusing to create block: pending blocks [%d] reached limit [%d]",
-			m.pendingHeaders.numPendingHeaders(), m.config.Node.MaxPendingBlocks)
-	}
-
-	var (
-		lastSignature  *types.Signature
-		lastHeaderHash types.Hash
-		lastDataHash   types.Hash
-		lastHeaderTime time.Time
-		err            error
-	)
-	height := m.store.Height()
-	newHeight := height + 1
-	// this is a special case, when first block is produced - there is no previous commit
-	if newHeight <= m.genesis.InitialHeight {
-		// Special handling for genesis block
-		lastSignature = &types.Signature{}
-	} else {
-		lastSignature, err = m.store.GetSignature(ctx, height)
-		if err != nil {
-			return fmt.Errorf("error while loading last commit: %w", err)
-		}
-		lastHeader, lastData, err := m.store.GetBlockData(ctx, height)
-		if err != nil {
-			return fmt.Errorf("error while loading last block: %w", err)
-		}
-		lastHeaderHash = lastHeader.Hash()
-		lastDataHash = lastData.Hash()
-		lastHeaderTime = lastHeader.Time()
-	}
-
-	var (
-		header    *types.SignedHeader
-		data      *types.Data
-		signature types.Signature
-	)
-
-	// Check if there's an already stored block at a newer height
-	// If there is use that instead of creating a new block
-	pendingHeader, pendingData, err := m.store.GetBlockData(ctx, newHeight)
-	if err == nil {
-		m.logger.Info("Using pending block", "height", newHeight)
-		header = pendingHeader
-		data = pendingData
-	} else {
-
-		execTxs, err := m.exec.GetTxs(ctx)
-		if err != nil {
-			m.logger.Error("failed to get txs from executor", "err", err)
-			// Continue but log the state
-			m.logger.Info("Current state",
-				"height", height,
-				"isProposer", m.isProposer,
-				"pendingHeaders", m.pendingHeaders.numPendingHeaders())
-		}
-
-		m.logger.Debug("Submitting transaction to sequencer",
-			"txCount", len(execTxs))
-		_, err = m.sequencer.SubmitRollupBatchTxs(ctx, coresequencer.SubmitRollupBatchTxsRequest{
-			RollupId: sequencing.RollupId(m.genesis.ChainID),
-			Batch:    &coresequencer.Batch{Transactions: execTxs},
-		})
-		if err != nil {
-			m.logger.Error("failed to submit rollup transaction to sequencer",
-				"err", err,
-				"chainID", m.genesis.ChainID)
-			// Add retry logic or proper error handling
-
-			m.logger.Debug("Successfully submitted transaction to sequencer")
-		}
-
-		txs, timestamp, err := m.getTxsFromBatch()
-		if errors.Is(err, ErrNoBatch) {
-			m.logger.Debug("No batch available, creating empty block")
-			// Create an empty block instead of returning
-			txs = [][]byte{}
-			timestamp = &time.Time{}
-			*timestamp = time.Now()
-		} else if err != nil {
-			return fmt.Errorf("failed to get transactions from batch: %w", err)
-		}
-		// sanity check timestamp for monotonically increasing
-		if timestamp.Before(lastHeaderTime) {
-			return fmt.Errorf("timestamp is not monotonically increasing: %s < %s", timestamp, m.getLastBlockTime())
-		}
-		m.logger.Info("Creating and publishing block", "height", newHeight)
-		header, data, err = m.createBlock(ctx, newHeight, lastSignature, lastHeaderHash, txs, *timestamp)
-		if err != nil {
-			return err
-		}
-		m.logger.Debug("block info", "num_tx", len(data.Txs))
-
-		/*
-		   here we set the SignedHeader.DataHash, and SignedHeader.Signature as a hack
-		   to make the block pass ValidateBasic() when it gets called by applyBlock on line 681
-		   these values get overridden on lines 687-698 after we obtain the IntermediateStateRoots.
-		*/
-		header.DataHash = data.Hash()
-		//header.Validators = m.getLastStateValidators()
-		//header.ValidatorHash = header.Validators.Hash()
-
-		signature, err = m.getSignature(header.Header)
-		if err != nil {
-			return err
-		}
-
-		// set the signature to current block's signed header
-		header.Signature = signature
-		err = m.store.SaveBlockData(ctx, header, data, &signature)
-		if err != nil {
-			return SaveBlockError{err}
-		}
-	}
-
-	newState, responses, err := m.applyBlock(ctx, header, data)
-	if err != nil {
-		if ctx.Err() != nil {
-			return err
-		}
-		// if call to applyBlock fails, we halt the node, see https://github.com/cometbft/cometbft/pull/496
-		panic(err)
-	}
-	// Before taking the hash, we need updated ISRs, hence after ApplyBlock
-	header.Header.DataHash = data.Hash()
-
-	signature, err = m.getSignature(header.Header)
-	if err != nil {
-		return err
-	}
-
-	// set the signature to current block's signed header
-	header.Signature = signature
-
-	// append metadata to Data before validating and saving
-	data.Metadata = &types.Metadata{
-		ChainID:      header.ChainID(),
-		Height:       header.Height(),
-		Time:         header.BaseHeader.Time,
-		LastDataHash: lastDataHash,
-	}
-	// Validate the created block before storing
-	if err := m.execValidate(m.lastState, header, data); err != nil {
-		return fmt.Errorf("failed to validate block: %w", err)
-	}
-
-	headerHeight := header.Height()
-
-	headerHash := header.Hash().String()
-	m.headerCache.setSeen(headerHash)
-
-	// SaveBlock commits the DB tx
-	err = m.store.SaveBlockData(ctx, header, data, &signature)
-	if err != nil {
-		return SaveBlockError{err}
-	}
-
-	// Commit the new state and block which writes to disk on the proxy app
-	appHash, err := m.execCommit(ctx, newState, header, data, responses)
-	if err != nil {
-		return err
-	}
-	// Update app hash in state
-	newState.AppHash = appHash
-
-	// SaveBlockResponses commits the DB tx
-	//err = m.store.SaveBlockResponses(ctx, headerHeight, responses)
-	//if err != nil {
-	//	return SaveBlockResponsesError{err}
-	//}
-
-	// Update the store height before submitting to the DA layer but after committing to the DB
-	m.store.SetHeight(ctx, headerHeight)
-
-	newState.DAHeight = atomic.LoadUint64(&m.daHeight)
-	// After this call m.lastState is the NEW state returned from ApplyBlock
-	// updateState also commits the DB tx
-	m.logger.Debug("updating state", "newState", newState)
-	err = m.updateState(ctx, newState)
-	if err != nil {
-		return err
-	}
-	m.recordMetrics(data)
-	// Check for shut down event prior to sending the header and block to
-	// their respective channels. The reason for checking for the shutdown
-	// event separately is due to the inconsistent nature of the select
-	// statement when multiple cases are satisfied.
-	select {
-	case <-ctx.Done():
-		return fmt.Errorf("unable to send header and block, context done: %w", ctx.Err())
-	default:
-	}
-
-	// Publish header to channel so that header exchange service can broadcast
-	m.HeaderCh <- header
-
-	// Publish block to channel so that block exchange service can broadcast
-	m.DataCh <- data
-
-	m.logger.Debug("successfully proposed header", "proposer", hex.EncodeToString(header.ProposerAddress), "height", headerHeight)
-
-	return nil
 }
 
 func (m *Manager) sign(payload []byte) ([]byte, error) {
@@ -1322,119 +516,6 @@ func (m *Manager) recordMetrics(data *types.Data) {
 	m.metrics.TotalTxs.Add(float64(len(data.Txs)))
 	m.metrics.BlockSizeBytes.Set(float64(data.Size()))
 	m.metrics.CommittedHeight.Set(float64(data.Metadata.Height))
-}
-func (m *Manager) submitHeadersToDA(ctx context.Context) error {
-	submittedAllHeaders := false
-	var backoff time.Duration
-	headersToSubmit, err := m.pendingHeaders.getPendingHeaders(ctx)
-	if len(headersToSubmit) == 0 {
-		// There are no pending headers; return because there's nothing to do, but:
-		// - it might be caused by error, then err != nil
-		// - all pending headers are processed, then err == nil
-		// whatever the reason, error information is propagated correctly to the caller
-		return err
-	}
-	if err != nil {
-		// There are some pending blocks but also an error. It's very unlikely case - probably some error while reading
-		// headers from the store.
-		// The error is logged and normal processing of pending blocks continues.
-		m.logger.Error("error while fetching blocks pending DA", "err", err)
-	}
-	numSubmittedHeaders := 0
-	attempt := 0
-	maxBlobSize, err := m.dalc.MaxBlobSize(ctx)
-	if err != nil {
-		return err
-	}
-	initialMaxBlobSize := maxBlobSize
-	gasPrice := m.gasPrice
-	initialGasPrice := gasPrice
-
-daSubmitRetryLoop:
-	for !submittedAllHeaders && attempt < maxSubmitAttempts {
-		select {
-		case <-ctx.Done():
-			break daSubmitRetryLoop
-		case <-time.After(backoff):
-		}
-
-		headersBz := make([][]byte, len(headersToSubmit))
-		for i, header := range headersToSubmit {
-			headerPb, err := header.ToProto()
-			if err != nil {
-				// do we drop the header from attempting to be submitted?
-				return fmt.Errorf("failed to transform header to proto: %w", err)
-			}
-			headersBz[i], err = headerPb.Marshal()
-			if err != nil {
-				// do we drop the header from attempting to be submitted?
-				return fmt.Errorf("failed to marshal header: %w", err)
-			}
-		}
-
-		ctx, cancel := context.WithTimeout(ctx, 60*time.Second) //TODO: make this configurable
-		defer cancel()
-		res := m.dalc.SubmitHeaders(ctx, headersBz, maxBlobSize, gasPrice)
-		switch res.Code {
-		case coreda.StatusSuccess:
-			m.logger.Info("successfully submitted Rollkit headers to DA layer", "gasPrice", gasPrice, "daHeight", res.DAHeight, "headerCount", res.SubmittedCount)
-			if res.SubmittedCount == uint64(len(headersToSubmit)) {
-				submittedAllHeaders = true
-			}
-			submittedBlocks, notSubmittedBlocks := headersToSubmit[:res.SubmittedCount], headersToSubmit[res.SubmittedCount:]
-			numSubmittedHeaders += len(submittedBlocks)
-			for _, block := range submittedBlocks {
-				m.headerCache.setDAIncluded(block.Hash().String())
-				err = m.setDAIncludedHeight(ctx, block.Height())
-				if err != nil {
-					return err
-				}
-			}
-			lastSubmittedHeight := uint64(0)
-			if l := len(submittedBlocks); l > 0 {
-				lastSubmittedHeight = submittedBlocks[l-1].Height()
-			}
-			m.pendingHeaders.setLastSubmittedHeight(ctx, lastSubmittedHeight)
-			headersToSubmit = notSubmittedBlocks
-			// reset submission options when successful
-			// scale back gasPrice gradually
-			backoff = 0
-			maxBlobSize = initialMaxBlobSize
-			if m.gasMultiplier > 0 && gasPrice != -1 {
-				gasPrice = gasPrice / m.gasMultiplier
-				if gasPrice < initialGasPrice {
-					gasPrice = initialGasPrice
-				}
-			}
-			m.logger.Debug("resetting DA layer submission options", "backoff", backoff, "gasPrice", gasPrice, "maxBlobSize", maxBlobSize)
-		case coreda.StatusNotIncludedInBlock, coreda.StatusAlreadyInMempool:
-			m.logger.Error("DA layer submission failed", "error", res.Message, "attempt", attempt)
-			backoff = m.config.DA.BlockTime * time.Duration(m.config.DA.MempoolTTL) //nolint:gosec
-			if m.gasMultiplier > 0 && gasPrice != -1 {
-				gasPrice = gasPrice * m.gasMultiplier
-			}
-			m.logger.Info("retrying DA layer submission with", "backoff", backoff, "gasPrice", gasPrice, "maxBlobSize", maxBlobSize)
-
-		case coreda.StatusTooBig:
-			maxBlobSize = maxBlobSize / 4
-			fallthrough
-		default:
-			m.logger.Error("DA layer submission failed", "error", res.Message, "attempt", attempt)
-			backoff = m.exponentialBackoff(backoff)
-		}
-
-		attempt += 1
-	}
-
-	if !submittedAllHeaders {
-		return fmt.Errorf(
-			"failed to submit all blocks to DA layer, submitted %d blocks (%d left) after %d attempts",
-			numSubmittedHeaders,
-			len(headersToSubmit),
-			attempt,
-		)
-	}
-	return nil
 }
 
 func (m *Manager) exponentialBackoff(backoff time.Duration) time.Duration {
