+++ block/transaction_aggregation.go
@@ -1,0 +1,99 @@
+package block
+
+import (
+	"context"
+	"time"
+)
+
+// AggregationLoop is responsible for aggregating transactions into rollup-blocks.
+func (m *Manager) AggregationLoop(ctx context.Context) {
+	initialHeight := m.genesis.InitialHeight //nolint:gosec
+	height := m.store.Height()
+	var delay time.Duration
+
+	// TODO(tzdybal): double-check when https://github.com/celestiaorg/rollmint/issues/699 is resolved
+	if height < initialHeight {
+		delay = time.Until(m.genesis.GenesisTime)
+	} else {
+		lastBlockTime := m.getLastBlockTime()
+		delay = time.Until(lastBlockTime.Add(m.config.Node.BlockTime))
+	}
+
+	if delay > 0 {
+		m.logger.Info("Waiting to produce block", "delay", delay)
+		time.Sleep(delay)
+	}
+
+	// blockTimer is used to signal when to build a block based on the
+	// rollup block time. A timer is used so that the time to build a block
+	// can be taken into account.
+	blockTimer := time.NewTimer(0)
+	defer blockTimer.Stop()
+
+	// Lazy Aggregator mode.
+	// In Lazy Aggregator mode, blocks are built only when there are
+	// transactions or every LazyBlockTime.
+	if m.config.Node.LazyAggregator {
+		m.lazyAggregationLoop(ctx, blockTimer)
+		return
+	}
+
+	m.normalAggregationLoop(ctx, blockTimer)
+}
+
+func (m *Manager) lazyAggregationLoop(ctx context.Context, blockTimer *time.Timer) {
+	// start is used to track the start time of the block production period
+	start := time.Now()
+	// lazyTimer is used to signal when a block should be built in
+	// lazy mode to signal that the chain is still live during long
+	// periods of inactivity.
+	lazyTimer := time.NewTimer(0)
+	defer lazyTimer.Stop()
+
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		// the m.bq.notifyCh channel is signalled when batch becomes available in the batch queue
+		case _, ok := <-m.bq.notifyCh:
+			if ok && !m.buildingBlock {
+				// set the buildingBlock flag to prevent multiple calls to reset the time
+				m.buildingBlock = true
+				// Reset the block timer based on the block time.
+				blockTimer.Reset(m.getRemainingSleep(start))
+			}
+			continue
+		case <-lazyTimer.C:
+		case <-blockTimer.C:
+		}
+		// Define the start time for the block production period
+		start = time.Now()
+		if err := m.publishBlock(ctx); err != nil && ctx.Err() == nil {
+			m.logger.Error("error while publishing block", "error", err)
+		}
+		// unset the buildingBlocks flag
+		m.buildingBlock = false
+		// Reset the lazyTimer to produce a block even if there
+		// are no transactions as a way to signal that the chain
+		// is still live.
+		lazyTimer.Reset(m.getRemainingSleep(start))
+	}
+}
+
+func (m *Manager) normalAggregationLoop(ctx context.Context, blockTimer *time.Timer) {
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case <-blockTimer.C:
+			// Define the start time for the block production period
+			start := time.Now()
+			if err := m.publishBlock(ctx); err != nil && ctx.Err() == nil {
+				m.logger.Error("error while publishing block", "error", err)
+			}
+			// Reset the blockTimer to signal the next block production
+			// period based on the block time.
+			blockTimer.Reset(m.getRemainingSleep(start))
+		}
+	}
+}
